Update Problem15.java
Bit messy. 

Map() seemed to have problems using an array as a key, so I changed the int[] array to a letter combination for the numbers 0 - 19 (A-T) to store as a key.

Second problem I had was I was searching in the wrong order where I should have been starting at the bottom right corner, instead I started at the top left which terrible confusion for a while. All sorted now though