Create Problem51.java
This entire code somewhat relies on the idea that we are guessing there will be a 1 (started with 0 but got no results), but then that assumes that all the ones will be changed. If we really want to make sure there aren't any false positives, then we would have to cross reference that check for 1 method and check if the same results come up when we check for 2, 3, 4, etc.